Segregating into source code and headers
